[Step1 피드백 반영]refactor: Apply singleton pattern for Retrofit reuse
[RetrofitAPI.kt]
- Retrofit 재사용이 가능하도록 싱글톤패턴 적용
[SearchDataRepository.kt]
- 카카오맵 API 검색 결과 가공 역할만 하도록 수정
[SearchViewModel.kt]
- 'SearchDataRepository'에 검색어를 넘기고,
검색 결과를 넘겨받도록 수정
